+++ tests/conftest.py
@@ -13,8 +13,7 @@ from iredis.redis_grammar import get_command_grammar
 from iredis.exceptions import InvalidArguments
 from iredis.config import Config, config as global_config
 
-
-TIMEOUT = 2
+TIMEOUT = 5
 HISTORY_FILE = ".iredis_history"
 
 
@@ -91,19 +90,23 @@ def config():
 def cli():
     """Open iredis subprocess to test"""
     f = tempfile.TemporaryFile("w")
-    config_content = dedent(
-        """
+    config_content = dedent("""
         [main]
         log_location =
         warning = True
-        """
-    )
+        """)
     f.write(config_content)
     f.close()
-    env = os.environ
+    env = os.environ.copy()
     env["PROMPT_TOOLKIT_NO_CPR"] = "1"
+    env["TERM"] = "xterm-256color"
 
-    child = pexpect.spawn(f"iredis -n 15 --iredisrc {f.name}", timeout=TIMEOUT, env=env)
+    child = pexpect.spawn(
+        f"iredis -n 15 --iredisrc {f.name}",
+        timeout=TIMEOUT,
+        env=env,
+        dimensions=(40, 120),
+    )
     child.logfile_read = open("cli_test.log", "ab")
     child.expect(["https://github.com/laixintao/iredis/issues", "127.0.0.1"])
     yield child
@@ -114,13 +117,11 @@ def cli():
 def raw_cli():
     """Open iredis subprocess to test"""
     TEST_IREDISRC = "/tmp/.iredisrc.test"
-    config_content = dedent(
-        """
+    config_content = dedent("""
         [main]
         log_location =
         warning = True
-        """
-    )
+        """)
 
     with open(TEST_IREDISRC, "w+") as test_iredisrc:
         test_iredisrc.write(config_content)
@@ -137,13 +138,11 @@ def raw_cli():
 @pytest.fixture(scope="function")
 def cli_without_warning():
     f = tempfile.TemporaryFile("w")
-    config_content = dedent(
-        """
+    config_content = dedent("""
         [main]
         log_location = /tmp/iredis1.log
         warning = False
-        """
-    )
+        """)
     f.write(config_content)
     f.close()